update all feature

+++ core/src/main/java/com/yurry/core/data/local/entity/FavoriteGameEntity.kt
@@ -14,7 +14,7 @@ data class FavoriteGameEntity(
     var name: String,
     var released: String,
     var background: String,
-    var rating: Int,
+    var rating: Float,
     var ratingTop: Int,
     var metacritic: Int
 )
